#11 Move server to api module
