+++ app.go
@@ -8,8 +8,8 @@ import (
 	"github.com/go-chi/chi/v5/middleware"
 	"net/http"
 	"slices"
+	"spt-give-ui/backend/api"
 	"spt-give-ui/backend/models"
-	"spt-give-ui/backend/spt"
 	"spt-give-ui/components"
 )
 
@@ -69,7 +69,7 @@ func NewChiRouter(app *App) *chi.Mux {
 		port := r.FormValue(contextPort)
 		version := app.ctx.Value(appVersion).(string)
 
-		serverInfo, err := spt.ConnectToSptServer(host, port)
+		serverInfo, err := api.ConnectToSptServer(host, port)
 		if err != nil {
 			templ.Handler(components.ErrorConnection(err.Error(), version)).ServeHTTP(w, r)
 			return
@@ -83,7 +83,7 @@ func NewChiRouter(app *App) *chi.Mux {
 		app.ctx = context.WithValue(app.ctx, contextHost, host)
 		app.ctx = context.WithValue(app.ctx, contextPort, port)
 
-		profiles, err := spt.LoadProfiles(host, port)
+		profiles, err := api.LoadProfiles(host, port)
 		if err != nil {
 			templ.Handler(components.ErrorConnection(err.Error(), version)).ServeHTTP(w, r)
 			return
@@ -95,7 +95,7 @@ func NewChiRouter(app *App) *chi.Mux {
 		version := app.ctx.Value(appVersion).(string)
 		sessionId := chi.URLParam(r, "id")
 		app.ctx = context.WithValue(app.ctx, contextSessionId, sessionId)
-		allItems, err := spt.LoadItems(app.ctx.Value(contextHost).(string), app.ctx.Value(contextPort).(string))
+		allItems, err := api.LoadItems(app.ctx.Value(contextHost).(string), app.ctx.Value(contextPort).(string))
 		if err != nil {
 			// TODO create new type of error template
 			templ.Handler(components.ErrorConnection(err.Error(), version)).ServeHTTP(w, r)
@@ -127,7 +127,7 @@ func NewChiRouter(app *App) *chi.Mux {
 		})
 		amount := allItems.Items[itemIdx].MaxStock
 
-		spt.AddItem(host, port, sessionId, itemId, amount)
+		api.AddItem(host, port, sessionId, itemId, amount)
 	})
 
 	return r
